requirements.txt
PyQt6>=6.0.0
Pillow>=9.0.0
requests>=2.25.0
openai>=1.0.0
google-generativeai>=0.3.0
paddleocr>=2.5.0
google-cloud-vision>=3.0.0
numpy>=1.20.0